Decide on using pf_scaffold
