+++ robot_localizer/scripts/pf_skeleton.py
@@ -1,0 +1,57 @@
+#!/usr/bin/env python3
+
+""" This is the starter code for the robot localization project """
+
+import rospy
+from geometry_msgs.msg import PoseWithCovarianceStamped, PoseArray, Pose
+
+from helper_functions import TFHelper
+from occupancy_field import OccupancyField
+
+
+class ParticleFilter(object):
+    """ The class that represents a Particle Filter ROS Node
+    """
+    def __init__(self):
+        rospy.init_node('pf')
+
+        # pose_listener responds to selection of a new approximate robot
+        # location (for instance using rviz)
+        rospy.Subscriber("initialpose",
+                         PoseWithCovarianceStamped,
+                         self.update_initial_pose)
+
+        # publisher for the particle cloud for visualizing in rviz.
+        self.particle_pub = rospy.Publisher("particlecloud",
+                                            PoseArray,
+                                            queue_size=10)
+
+        # create instances of two helper objects that are provided to you
+        # as part of the project
+        self.occupancy_field = OccupancyField()
+        self.transform_helper = TFHelper()
+
+    def update_initial_pose(self, msg):
+        """ Callback function to handle re-initializing the particle filter
+            based on a pose estimate.  These pose estimates could be generated
+            by another ROS Node or could come from the rviz GUI """
+        xy_theta = \
+            self.transform_helper.convert_pose_to_xy_and_theta(msg.pose.pose)
+
+        # initialize your particle filter based on the xy_theta tuple
+
+        # Use the helper functions to fix the transform
+
+    def run(self):
+        r = rospy.Rate(5)
+
+        while not(rospy.is_shutdown()):
+            # in the main loop all we do is continuously broadcast the latest
+            # map to odom transform
+            self.transform_helper.send_last_map_to_odom_transform()
+            r.sleep()
+
+
+if __name__ == '__main__':
+    n = ParticleFilter()
+    n.run()
